// src/bot/common/enums/set-game.enum.ts
export enum GameMap {

  // DEFAULT GAMES MAP (NO DLC)
  TRANSIT = "transit",
  FARM = "farm",
  TOWN = "town",
  DINER = "diner",
  BUS_DEPOT = "bus_depot",

  // DLC MAPS
  ORIGINS = "origins",
  NUKETOWN = "nuketown",
  MOB_OF_THE_DEAD = "mob_of_the_dead",
  DIE_RISE = "die_rise",
  BURIED = "buried"
}

export enum GameMode {
  CLASSIC = "classic",
  STANDARD = "standard",
  GRIEF = "grief",
  TURNED = "cleansed", // Same as CLEANSED
}